+++ etc/supervisor_available/cron_db.conf
@@ -1,5 +1,5 @@
 [program:cron_db]
-command = /app/bin/cron_db.py
+command = /app/bin/cron.py db
 numprocs = 1
 autostart = true
 user = webdev
